requirements.txt
allure-python-commons~=2.13.5
PyAutoGUI~=0.9.54
pytesseract~=0.3.13
opencv-python~=4.10.0.84
pytest~=8.3.3
PyGetWindow~=0.0.9